Move CBMA methods up a level.
This removes the CBMA submodule to simplify the organization of the
package.
