+++ examples/02_meta-analyses/cbma.py
@@ -36,7 +36,7 @@ mask_img = dset.masker.mask_img
 ###############################################################################
 # MKDA density analysis
 # --------------------------------------------------
-mkda = nimare.meta.cbma.mkda.MKDADensity(kernel__r=10)
+mkda = nimare.meta.mkda.MKDADensity(kernel__r=10)
 mkda.fit(dset)
 corr = nimare.correct.FWECorrector(method='montecarlo', n_iters=10, n_cores=1)
 cres = corr.transform(mkda.results)
@@ -46,7 +46,7 @@ plot_stat_map(cres.get_map('logp_level-voxel_corr-FWE_method-montecarlo'),
 ###############################################################################
 # MKDA Chi2 with FDR correction
 # --------------------------------------------------
-mkda = nimare.meta.cbma.mkda.MKDAChi2(kernel__r=10)
+mkda = nimare.meta.mkda.MKDAChi2(kernel__r=10)
 dset1 = dset.slice(dset.ids)
 dset2 = dset.slice(dset.ids)
 mkda.fit(dset1, dset2)
@@ -70,7 +70,7 @@ plot_stat_map(cres.get_map('z_desc-consistency_level-voxel_corr-FWE_method-monte
 ###############################################################################
 # KDA
 # --------------------------------------------------
-kda = nimare.meta.cbma.mkda.KDA(kernel__r=10)
+kda = nimare.meta.mkda.KDA(kernel__r=10)
 kda.fit(dset)
 corr = nimare.correct.FWECorrector(method='montecarlo', n_iters=10, n_cores=1)
 cres = corr.transform(kda.results)
@@ -80,7 +80,7 @@ plot_stat_map(cres.get_map('logp_level-voxel_corr-FWE_method-montecarlo'),
 ###############################################################################
 # ALE
 # --------------------------------------------------
-ale = nimare.meta.cbma.ale.ALE()
+ale = nimare.meta.ale.ALE()
 ale.fit(dset)
 corr = nimare.correct.FWECorrector(method='montecarlo', n_iters=10, n_cores=1)
 cres = corr.transform(ale.results)
@@ -91,7 +91,7 @@ plot_stat_map(cres.get_map('logp_level-cluster_corr-FWE_method-montecarlo'),
 # SCALE
 # --------------------------------------------------
 ijk = np.vstack(np.where(mask_img.get_fdata())).T
-scale = nimare.meta.cbma.ale.SCALE(ijk=ijk, n_iters=10, n_cores=1)
+scale = nimare.meta.ale.SCALE(ijk=ijk, n_iters=10, n_cores=1)
 scale.fit(dset)
 plot_stat_map(scale.results.get_map('z'), cut_coords=[0, 0, -8],
               draw_cross=False, cmap='RdBu_r')
