+++ nimare/meta/__init__.py
@@ -2,7 +2,6 @@
 Coordinate-, image-, and effect-size-based meta-analysis estimators.
 """
 
-from . import cbma
-from . import ibma
+from . import ale, mkda, kernel, ibma
 
-__all__ = ['cbma', 'ibma']
+__all__ = ['ale', 'mkda', 'kernel', 'ibma']
