--- nimare/meta/cbma/__init__.py
@@ -1,10 +1,0 @@
-"""
-Meta-analysis estimators
-"""
-
-from .ale import ALE, SCALE, ALESubtraction
-from .mkda import MKDAChi2, MKDADensity, KDA
-from . import kernel
-
-__all__ = ['ALE', 'SCALE', 'ALESubtraction', 'MKDAChi2', 'MKDADensity', 'KDA',
-           'kernel']
