+++ nimare/meta/kernel.py
@@ -10,9 +10,9 @@ import nibabel as nib
 
 from nilearn.image import resample_to_img, math_img
 from .utils import compute_ma, get_ale_kernel, peaks2maps
-from ...transforms import vox2mm
+from ..transforms import vox2mm
 
-from ...base import KernelTransformer
+from ..base import KernelTransformer
 
 
 class ALEKernel(KernelTransformer):
