+++ nimare/meta/mkda.py
@@ -11,12 +11,12 @@ from scipy import ndimage, special
 from statsmodels.sandbox.stats.multicomp import multipletests
 
 from .kernel import MKDAKernel, KDAKernel
-from ...results import MetaResult
-from ...base import CBMAEstimator
-from ...stats import null_to_p, one_way, two_way
-from ...transforms import p_to_z
-from ...due import due
-from ... import references
+from ..results import MetaResult
+from ..base import CBMAEstimator
+from ..stats import null_to_p, one_way, two_way
+from ..transforms import p_to_z
+from ..due import due
+from .. import references
 
 LGR = logging.getLogger(__name__)
