+++ nimare/tests/test_meta_kernel.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 from scipy.ndimage.measurements import center_of_mass
 
-from nimare.meta.cbma import kernel
+from nimare.meta import kernel
 from nimare.utils import get_template, get_masker
