+++ nimare/tests/test_meta_mkda.py
@@ -1,10 +1,10 @@
 """
-Test nimare.meta.cbma.mkda (KDA-based meta-analytic algorithms).
+Test nimare.meta.mkda (KDA-based meta-analytic algorithms).
 """
 import pytest
 
 import nimare
-from nimare.meta.cbma import mkda, kernel
+from nimare.meta import mkda, kernel
 from nimare.correct import FWECorrector, FDRCorrector
